requirements.txt
flask
opencv-python
openvino-dev
speechrecognition
pyttsx3
g4f
